refactor: move packages around so its easier to see the routes
Make the routing package just used for setting up the routing, and move the routes out into their
own routes package.
