+++ src/main/kotlin/com/chesire/orochi/plugins/routing/Routing.kt
@@ -1,7 +1,7 @@
 package com.chesire.orochi.plugins.routing
 
-import com.chesire.orochi.plugins.routing.kitsu.kitsu
-import com.chesire.orochi.plugins.routing.status.status
+import com.chesire.orochi.routes.kitsu.kitsu
+import com.chesire.orochi.routes.status.status
 import io.ktor.application.Application
 import io.ktor.routing.routing
